Move issue_tracker functionality into package
Move issue_tracker.py file into issue_tracker package
and rename it into assessment_functionality.
This moving is needed because we will store in issue_tracker
package functionality for integration not only for assessments but
for example for issue object.
