+++ src/ggrc/models/hooks/issue_tracker/assessment_integration.py
